Implemented websocket for Catalyst, but is broken
Turns out there doesn't seem to be any way to make WebSockets
work in Catalyst without breaking the API.  This implementation
should theoretically work, but a default plack middleware is
inserting an empty body due to the status code being 101 because
101 has no response.  This breaks the write_fh method which was
trying to pass back 2 arguments to the server.  That seems like
a bug that should get fixed, so I'm committing this version
before I mess it up with ugly hacks.